Refactor code and extract handlers into seperate file
