+++ app/javascript/react/locales/en/translation.json
@@ -51,6 +51,7 @@
     "calendarTitle": "Measurements calendar"
   },
 <<<<<<< HEAD
+<<<<<<< HEAD
   "headerToggle": {
     "arrowIcon": "Arrow Icon"
 =======
@@ -58,5 +59,10 @@
     "emptyInputMessage": "Please enter a valid number"
 
 >>>>>>> 60fa348a (Fix inputs behaviours)
+=======
+  "thresholdConfigurator": {
+    "emptyInputMessage": "Please enter a valid number",
+    "validValueMessage":  "Value must be between {{minValue}} and {{maxValue}}"
+>>>>>>> 60f771e0 (Refactor code and extract handlers into seperate file)
   }
 }
